Updated script according to ARs

+++ scripts/update_devices_db.py
@@ -190,7 +190,7 @@ async def upload_release_asset(session, token, repo, asset_path, release_version
     print(f"Upload completed for: {os.path.basename(asset_path)}.")
     return result
 
-def setSDKToDeviceInDb(dbs, regex, ai_sdk):
+def setSDKSupport(dbs, regex, ai_sdk):
     for eachDb in dbs:
         if eachDb:
             updateTableCollumn(
@@ -216,21 +216,31 @@ def setSDKToDeviceInDb(dbs, regex, ai_sdk):
 
     return
 
-def removeDeviceFromDb(dbs, regex):
+def removeDeviceFromDb(dbs, regex, delete_device):
     for eachDb in dbs:
         if eachDb:
-            deleteFromTable(
-                eachDb,
-                f'''
-                DELETE FROM Devices
-                WHERE uid REGEXP "{regex}";
-                '''
-            )
+            if delete_device:
+                deleteFromTable(
+                    eachDb,
+                    f'''
+                    DELETE FROM Devices
+                    WHERE uid REGEXP "{regex}";
+                    '''
+                )
+            else:
+                updateTableCollumn(
+                    eachDb,
+                    "Devices",
+                    "sdk_support",
+                    0,
+                    "uid",
+                    regex
+                )
 
     return
 
 ## Main runner
-async def main(token, repo, index="Test", action="Set SDKToDevice", regex="", ai_sdk=False):
+async def main(token, repo, index="Test", action="Set sdk_support", regex="", delete_device=False, ai_sdk=False):
     ## Download the database first
     ## Always use latest release
     dbName = 'necto_db_dev'
@@ -245,10 +255,10 @@ async def main(token, repo, index="Test", action="Set SDKToDevice", regex="", ai
     )
 
     ## Update database with requested settings
-    if "Set SDKToDevice" == action:
-        setSDKToDeviceInDb([databaseErp, databaseNecto], regex, ai_sdk)
+    if "Set sdk_support" == action:
+        setSDKSupport([databaseErp, databaseNecto], regex, ai_sdk)
     else:
-        removeDeviceFromDb([databaseErp, databaseNecto], regex)
+        removeDeviceFromDb([databaseErp, databaseNecto], regex, delete_device)
 
     ## Reupload databases over existing assets
     archive_path = compress_directory_7z(os.path.join(os.path.dirname(__file__), 'databases'), f'{dbPackageName}.7z')
@@ -258,7 +268,7 @@ async def main(token, repo, index="Test", action="Set SDKToDevice", regex="", ai
         async with aiohttp.ClientSession() as session:
             upload_result = await upload_release_asset(session, token, repo, databaseErp, release_version)
 
-    ## Step 16 - overwrite the existing necto_db.db in root with newly generated one
+    ## Overwrite the existing necto_db.db in root with newly generated one
     shutil.copy2(databaseNecto, os.path.join(os.getcwd(), f'{dbName}.db'))
     ## ------------------------------------------------------------------------------------ ##
 ## EOF Main runner
@@ -280,8 +290,9 @@ if __name__ == "__main__":
     parser.add_argument("token", help="GitHub Token")
     parser.add_argument("repo", help="Repository name, e.g., 'username/repo'")
     parser.add_argument('index', type=str, help='Index selection - Live/Test.', default="Test")
-    parser.add_argument('action', type=str, help='Action selection - Remove/Update Devices.', default="Set SDKToDevice")
+    parser.add_argument('action', type=str, help='Action selection - Remove/Update Devices.', default="Set sdk_support")
     parser.add_argument('regex', type=str, help='Regex for Devices that need to be updated.', default="")
+    parser.add_argument('delete_device', type=str2bool, help='If True - will remove device from DB completely.', default=False)
     parser.add_argument('--ai_sdk', type=str2bool, help='If True - will add AI_GENERATED field to sdk_config.', default=False)
 
     ## Parse the arguments
@@ -292,6 +303,7 @@ if __name__ == "__main__":
         main(
             args.token, args.repo,
             args.index, args.action,
-            args.regex, args.ai_sdk
+            args.regex, args.delete_device,
+            args.ai_sdk
         )
     )
